+++ test/access-assignment.spec.js
@@ -6,7 +6,7 @@ describe('User and role access assignment:', () => {
     testHelper.initSyncFunction('build/sync-functions/test-access-assignment-sync-function.js');
   });
 
-  describe('Static assignment of channels to users and roles', () => {
+  describe('Static assignment of channels to users and roles and assignment of roles to users', () => {
     const expectedStaticAssignments = [
       {
         expectedType: 'channel',
@@ -76,9 +76,9 @@ describe('User and role access assignment:', () => {
     });
   });
 
-  describe('Dynamic assignment of channels to users and roles', () => {
+  describe('Dynamic definition of access assignments constraint', () => {
     const doc = {
-      _id: 'dynamicAccessDoc',
+      _id: 'dynamicAccessConstraintDoc',
       users: [ 'user1', 'user2' ],
       roles: [ 'role1', 'role2' ]
     };
@@ -112,7 +112,7 @@ describe('User and role access assignment:', () => {
 
     it('is applied when replacing a deleted document', () => {
       const oldDoc = {
-        _id: 'dynamicAccessDoc',
+        _id: 'dynamicAccessConstraintDoc',
         _deleted: true
       };
 
@@ -123,13 +123,13 @@ describe('User and role access assignment:', () => {
     });
 
     it('is applied when replacing an existing valid document', () => {
-      const oldDoc = { _id: 'dynamicAccessDoc' };
+      const oldDoc = { _id: 'dynamicAccessConstraintDoc' };
 
       testHelper.verifyDocumentReplaced(doc, oldDoc, 'write', expectedDynamicAssignments);
     });
 
     it('is applied when deleting an existing document', () => {
-      const oldDoc = { _id: 'dynamicAccessDoc' };
+      const oldDoc = { _id: 'dynamicAccessConstraintDoc' };
 
       const expectedDeleteAssignments = [
         {
@@ -160,7 +160,7 @@ describe('User and role access assignment:', () => {
 
     it('is NOT applied when creating an invalid document', () => {
       const doc = {
-        _id: 'dynamicAccessDoc',
+        _id: 'dynamicAccessConstraintDoc',
         users: [ 'user1' ],
         roles: [ 'role1' ],
         invalidProperty: 'foobar'
@@ -174,12 +174,124 @@ describe('User and role access assignment:', () => {
 
     it('is NOT applied when replacing an invalid document', () => {
       const doc = {
-        _id: 'dynamicAccessDoc',
+        _id: 'dynamicAccessConstraintDoc',
         users: [ 'user1' ],
         roles: [ 'role1' ],
         invalidProperty: 'foobar'
       };
-      const oldDoc = { _id: 'dynamicAccessDoc' };
+      const oldDoc = { _id: 'dynamicAccessConstraintDoc' };
+
+      expect(() => {
+        testHelper.syncFunction(doc, oldDoc);
+      }).to.throw();
+      expect(testHelper.access.callCount).to.equal(0);
+    });
+  });
+
+  describe('Dynamic definition of each assignment entry\'s channels, users and roles', () => {
+    const doc = {
+      _id: 'dynamicAccessEntryItemsDoc',
+      users: [ 'user1', 'user2' ],
+      roles: [ 'role1', 'role2' ]
+    };
+    const expectedDynamicAssignments = [
+      {
+        expectedType: 'channel',
+        expectedUsers: doc.users,
+        expectedRoles: doc.roles,
+        expectedChannels: [ `${doc._id}-channel1`, `${doc._id}-channel2` ]
+      },
+      {
+        expectedType: 'channel',
+        expectedUsers: doc.users,
+        expectedChannels: [ `${doc._id}-channel3` ]
+      },
+      {
+        expectedType: 'channel',
+        expectedRoles: doc.roles,
+        expectedChannels: [ `${doc._id}-channel4` ]
+      },
+      {
+        expectedType: 'role',
+        expectedUsers: doc.users,
+        expectedRoles: doc.roles
+      }
+    ];
+
+    it('is applied when creating a valid document', () => {
+      testHelper.verifyDocumentCreated(doc, 'write', expectedDynamicAssignments);
+    });
+
+    it('is applied when replacing a deleted document', () => {
+      const oldDoc = {
+        _id: 'dynamicAccessEntryItemsDoc',
+        _deleted: true
+      };
+
+      // The access assignment functions for this document type are set up to return different values if they receive an oldDoc parameter
+      // that has _deleted set to true. However, that should never happen because the sync function template is supposed to replace such
+      // cases with a null value. This test verifies that replacement occurs as expected.
+      testHelper.verifyDocumentAccepted(doc, oldDoc, 'write', expectedDynamicAssignments);
+    });
+
+    it('is applied when replacing an existing valid document', () => {
+      const oldDoc = { _id: 'dynamicAccessEntryItemsDoc' };
+
+      testHelper.verifyDocumentReplaced(doc, oldDoc, 'write', expectedDynamicAssignments);
+    });
+
+    it('is applied when deleting an existing document', () => {
+      const oldDoc = { _id: 'dynamicAccessEntryItemsDoc' };
+
+      const expectedDeleteAssignments = [
+        {
+          expectedType: 'channel',
+          expectedUsers: null,
+          expectedChannels: [ `${doc._id}-channel3` ]
+        },
+        {
+          expectedType: 'channel',
+          expectedRoles: null,
+          expectedChannels: [ `${doc._id}-channel4` ]
+        },
+        {
+          expectedType: 'channel',
+          expectedUsers: null,
+          expectedRoles: null,
+          expectedChannels: [ `${doc._id}-channel2`, `${doc._id}-channel1` ]
+        },
+        {
+          expectedType: 'role',
+          expectedUsers: null,
+          expectedRoles: null
+        }
+      ];
+
+      testHelper.verifyDocumentDeleted(oldDoc, 'write', expectedDeleteAssignments);
+    });
+
+    it('is NOT applied when creating an invalid document', () => {
+      const doc = {
+        _id: 'dynamicAccessEntryItemsDoc',
+        users: [ 'user1' ],
+        roles: [ 'role1' ],
+        invalidProperty: 'foobar'
+      };
+
+      expect(() => {
+        testHelper.syncFunction(doc);
+      }).to.throw();
+      expect(testHelper.access.callCount).to.equal(0);
+    });
+
+    it('is NOT applied when replacing an invalid document', () => {
+      const doc = {
+        _id: 'dynamicAccessEntryItemsDoc',
+        users: [ 'user1' ],
+        roles: [ 'role1' ],
+        invalidProperty: 'foobar'
+      };
+      const oldDoc = { _id: 'dynamicAccessEntryItemsDoc' };
 
       expect(() => {
         testHelper.syncFunction(doc, oldDoc);