+++ test/resources/access-assignment-doc-definitions.js
@@ -29,11 +29,11 @@
       }
     ]
   },
-  dynamicAccessDoc: {
-    channels: { write: 'write' },
-    typeFilter: function(doc) {
-      return doc._id === 'dynamicAccessDoc';
+  dynamicAccessEntryItemsDoc: {
+    typeFilter: function(doc, oldDoc, docType) {
+      return doc._id === docType;
     },
+    channels: { write: 'write' },
     propertyValidators: {
       users: {
         type: 'array'
@@ -45,16 +45,12 @@
     accessAssignments: [
       {
         type: 'role',
-        users: function(doc, oldDoc) {
-          // The sync function template should automatically replace the oldDoc param value with null if its _deleted property is true
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
-        },
-        roles: function(doc, oldDoc) {
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.roles;
-        }
+        users: function(doc, oldDoc) { return doc.users; },
+        roles: function(doc, oldDoc) { return doc.roles; }
       },
       {
         type: 'channel',
+        // The sync function template should automatically replace the oldDoc param value with null if its _deleted property is true
         users: function(doc, oldDoc) {
           return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
         },
@@ -66,23 +62,57 @@
         }
       },
       {
-        // With no "type" property defined, it should default to channel access assignment
-        roles: function(doc, oldDoc) {
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.roles;
-        },
-        channels: function(doc, oldDoc) {
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc._id + '-channel4';
-        }
+        // With a null "type" property defined, it should default to channel access assignment
+        type: null,
+        roles: function(doc, oldDoc) { return doc.roles; },
+        channels: function(doc, oldDoc) { return doc._id + '-channel4'; }
       },
       {
         // With no "type" property defined, it should default to channel access assignment
-        users: function(doc, oldDoc) {
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users;
-        },
-        channels: function(doc, oldDoc) {
-          return (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc._id + '-channel3';
-        }
+        users: function(doc, oldDoc) { return doc.users; },
+        channels: function(doc, oldDoc) { return doc._id + '-channel3'; }
       }
     ]
+  },
+  dynamicAccessConstraintDoc: {
+    typeFilter: function(doc, oldDoc, docType) {
+      return doc._id === docType;
+    },
+    channels: { write: [ 'write' ] },
+    propertyValidators: {
+      users: {
+        type: 'array'
+      },
+      roles: {
+        type: 'array'
+      }
+    },
+    accessAssignments: function(doc, oldDoc) {
+      return [
+        {
+          type: 'role',
+          users: doc.users,
+          roles: doc.roles
+        },
+        {
+          type: 'channel',
+          // The sync function template should automatically replace the oldDoc param value with null if its _deleted property is true
+          users: (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.users,
+          roles: (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : doc.roles,
+          channels: (oldDoc && oldDoc._deleted) ? 'this-should-never-happen' : [ doc._id + '-channel1', doc._id + '-channel2' ]
+        },
+        {
+          // With no "type" property defined, it should default to channel access assignment
+          roles: doc.roles,
+          channels: doc._id + '-channel4'
+        },
+        {
+          // With a null "type" property defined, it should default to channel access assignment
+          type: null,
+          users: doc.users,
+          channels: doc._id + '-channel3'
+        }
+      ];
+    }
   }
 }